Implement test for config -> redoc json generation
