+++ deployment/src/test/resources/serializer/all-false/config.properties
@@ -1,0 +1,17 @@
+# All optional config values set to false/disabled values
+quarkus.redoc.path=redoc
+quarkus.redoc.title=API Documentation
+quarkus.redoc.always-include=false
+quarkus.redoc.hide-download-buttons=false
+quarkus.redoc.hide-schema-titles=false
+quarkus.redoc.json-samples-expand-level=2
+quarkus.redoc.max-displayed-enum-values=5
+quarkus.redoc.layout=three-panel
+quarkus.redoc.only-required-in-samples=false
+quarkus.redoc.sort-required-props-first=false
+quarkus.redoc.schemas-expansion-level=1
+quarkus.redoc.scroll-y-offset=0
+quarkus.redoc.show-extensions=false
+quarkus.redoc.sanitize=false
+quarkus.redoc.generated-samples-max-depth=4
+quarkus.redoc.hide-properties-prefix=false
