+++ deployment/src/test/resources/serializer/all-true/config.properties
@@ -1,0 +1,18 @@
+# All optional config values set to true/enabled values
+quarkus.redoc.path=redoc
+quarkus.redoc.title=API Documentation
+quarkus.redoc.always-include=true
+quarkus.redoc.hide-download-buttons=true
+quarkus.redoc.hide-schema-titles=true
+quarkus.redoc.json-samples-expand-level=all
+quarkus.redoc.max-displayed-enum-values=20
+quarkus.redoc.layout=stacked
+quarkus.redoc.only-required-in-samples=true
+quarkus.redoc.sort-required-props-first=true
+quarkus.redoc.schemas-expansion-level=all
+quarkus.redoc.scroll-y-offset=100
+quarkus.redoc.show-extensions=true
+quarkus.redoc.sanitize=true
+quarkus.redoc.schema-definitions-tag-name=Models
+quarkus.redoc.generated-samples-max-depth=10
+quarkus.redoc.hide-properties-prefix=true
