+++ deployment/src/test/resources/serializer/unset/config.properties
@@ -1,0 +1,4 @@
+# All optional config values are unset - only defaults
+quarkus.redoc.path=redoc
+quarkus.redoc.title=API Documentation
+quarkus.redoc.always-include=false
