+++ deployment/src/test/resources/serializer/unset/expected.json
@@ -1,0 +1,3 @@
+{
+  "schemaDefinitionsTagName": "Schemas"
+}
